4.0.3版本和ETC部分功能

+++ pickerview/src/main/java/com/bigkoo/pickerview/utils/ChinaDate.java
@@ -290,7 +290,7 @@ public class ChinaDate {
             return "二十";
         if (day == 30)
             return "三十";
-        int two = (int) ((day) / 10);
+        int two = (day) / 10;
         if (two == 0)
             a = "初";
         if (two == 1)
@@ -299,7 +299,7 @@ public class ChinaDate {
             a = "廿";
         if (two == 3)
             a = "三";
-        int one = (int) (day % 10);
+        int one = day % 10;
         switch (one) {
             case 1:
                 a += "一";
